requirements.txt
transformers==4.49.0
trl==0.15.2
peft==0.14.0
unsloth==2025.3.19
python-dotenv==1.1.0
hydra-core==1.3.2
pydantic==2.11.4